+++ tests/client/jwt_test.py
@@ -1,0 +1,13 @@
+"""
+Tests for the JWT endpoint.
+"""
+
+import pytest
+
+
+@pytest.mark.vcr
+def test_jwt(client):
+    """Tests fetching of JWT
+    """
+    resp = client.jwt()
+    assert "eyJ0eXAiOiJK" in resp.jwt